[Front] refacto
